+++ devel/fortio/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "grpcrunner.go",
         "http.go",
         "httprunner.go",
         "logger.go",
@@ -11,17 +10,12 @@ go_library(
         "stats.go",
     ],
     visibility = ["//visibility:public"],
-    deps = [
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//health/grpc_health_v1:go_default_library",
-    ],
 )
 
 go_test(
     name = "go_default_test",
     size = "small",
     srcs = [
-        "grpcrunner_test.go",
         "http_test.go",
         "httprunner_test.go",
         "logger_test.go",
@@ -29,9 +23,4 @@ go_test(
         "stats_test.go",
     ],
     library = ":go_default_library",
-    deps = [
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//health:go_default_library",
-        "@org_golang_google_grpc//health/grpc_health_v1:go_default_library",
-    ],
 )
