+++ devel/fortio/Dockerfile.echosrv
@@ -1,0 +1,10 @@
+# Build the binaries in larger image
+FROM golang:1.8.3 as build
+WORKDIR /go/src/istio.io/istio/devel
+COPY . fortio
+RUN CGO_ENABLED=0 GOOS=linux go build -a -ldflags '-s' -o echosrv.bin istio.io/istio/devel/fortio/cmd/echosrv
+# Minimal image with just the binary
+FROM scratch
+COPY --from=build /go/src/istio.io/istio/devel/echosrv.bin /usr/local/bin/echosrv
+EXPOSE 8080
+ENTRYPOINT ["/usr/local/bin/echosrv"]
