+++ devel/fortio/Dockerfile.fortio
@@ -1,0 +1,11 @@
+# Build the binaries in larger image
+FROM golang:1.8.3 as build
+WORKDIR /go/src/istio.io/istio/devel
+RUN go get google.golang.org/grpc
+COPY . fortio
+RUN CGO_ENABLED=0 GOOS=linux go build -a -ldflags '-s' -o fortio.bin istio.io/istio/devel/fortio/cmd/fortio
+# Minimal image with just the binary and certs
+FROM scratch
+COPY --from=build /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
+COPY --from=build /go/src/istio.io/istio/devel/fortio.bin /usr/local/bin/fortio
+ENTRYPOINT ["/usr/local/bin/fortio"]
